Before Migrating to float type

+++ Simple 2D Landscape/WinForms/Form1.cs
@@ -18,7 +18,7 @@ namespace Simple_2D_Landscape
 		{
 			InitializeComponent();
 
-            daseffect<short> test = new daseffect<short>();
+            daseffect<float> test = new daseffect<float>();
 
 
 		}
